fix compiler warnings / cosmetics

+++ player2/components/include/ring.h
@@ -59,8 +59,8 @@ public:
 
     virtual ~Ring_c( void ) {};
 
-    virtual RingStatus_t Insert(  unsigned int	 Value ) = 0;
-    virtual RingStatus_t Extract( unsigned int  *Value,
+    virtual RingStatus_t Insert(  uintptr_t Value ) = 0;
+    virtual RingStatus_t Extract( uintptr_t *Value,
                                   unsigned int   BlockingPeriod = RING_NONE_BLOCKING ) = 0;
     virtual RingStatus_t Flush( void ) = 0;
     virtual bool         NonEmpty( void ) = 0;
